+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 * Fix: False positive ad blocker detection on some devices
 * Fix: Negative speed issue
 * Fix: Restoring the notification icon on Windows after restarting Windows Explorer
+* Fix: Access to local network when IncludeLocalNetwork is enabled
 * Feature: Internal Ad Provider
 * Feature: Show more details in connection progress
+* Feature: Allow set custom DNS servers
 
 # v7.3.748
 * Fix: Minor UI issues and translations